[Refactor] controller 분리에 따른 DocsTest 수정 (#80)
